Use `attr.output` instead of deprecated `outputs` dict
Backwards compatibility is preserved by wrapping the `jar_jar` rule in a
macro that populates the new `output_jar` attribute if the user doesn't
set it. The output of a `jar_jar` target can still be referenced as a
file target as demonstrated by new tests.

In addition, user can now provide a custom name for the output jar.