Improve remapper cache
The previous remapper implementation contained some branches in which
the returned values weren't cached. Most importantly, when a name wasn't
mapped, the `null` value returned in that case is persisted as `null` in
the map and thus wasn't distinguishable from a cache miss with `get()`.
This is worked around by using `computeIfAbsent`.

Improves the JUnit benchmark's runtime by ~2%. The impact is larger when
a larger fraction of identifiers isn't mapped by rules.